Fixed Problem #10.

+++ Project Euler/src/Main.cpp
@@ -15,7 +15,7 @@ unsigned int Problem_6();
 unsigned int Problem_7();
 unsigned int Problem_8();
 unsigned int Problem_9();
-unsigned int Problem_10();
+long long int Problem_10();
 
 
 int main()
